requirements.txt
anyio==3.6.2
autopep8==2.0.1
bcrypt==4.0.1
boto3==1.26.71
botocore==1.29.71
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
docopt==0.6.2
fastapi==0.92.0
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
install==1.3.5
jmespath==1.0.1
passlib==1.7.4
pipreqs==0.4.11
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.4
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-magic==0.4.27
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarg==0.1.9
